gives dictionary and word classes discrete files
